refactor: move new angular component into specific folder
