+++ gravitee-apim-console-webui/src/management/instances/instance-details/instance-details-environment/instance-details-environment.component.scss
@@ -1,7 +1,7 @@
 @use 'sass:map';
 @use '@angular/material' as mat;
 @use '@gravitee/ui-particles-angular' as gio;
-@use '../../../../scss/gio-layout' as gio-layout;
+@use 'src/scss/gio-layout' as gio-layout;
 
 :host {
   @include gio-layout.gio-responsive-margin-container;
