+++ gravitee-apim-console-webui/src/management/instances/instance-details/instance-details-environment/instance-details-environment.component.spec.ts
