+++ gravitee-apim-console-webui/src/management/instances/instance-details/instance-details-monitoring/instance-details-monitoring.module.ts
