+++ gravitee-apim-console-webui/src/management/management.module.ts
@@ -27,8 +27,8 @@ import { EnvironmentApplicationModule } from './application/environment-applicat
 import { ApisModule } from './api/apis.module';
 import { SettingsNavigationModule } from './configuration/settings-navigation/settings-navigation.module';
 import { InstancesNavigationModule } from './instances/details/instances-navigation/instances-navigation.module';
-import { InstanceDetailsMonitoringModule } from './instances/details/instance-details-monitoring/instance-details-monitoring.module';
-import { InstanceDetailsEnvironmentModule } from './instances/details/instance-details-environment/instance-details-environment.module';
+import { InstanceDetailsMonitoringModule } from './instances/instance-details/instance-details-monitoring/instance-details-monitoring.module';
+import { InstanceDetailsEnvironmentModule } from './instances/instance-details/instance-details-environment/instance-details-environment.module';
 
 import { GioConfirmDialogModule } from '../shared/components/gio-confirm-dialog/gio-confirm-dialog.module';
 import { GioPermissionModule } from '../shared/components/gio-permission/gio-permission.module';
